requirements.txt
flask~=3.1.0
faiss-cpu~=1.9.0
langchain~=0.3.10
python-dotenv~=1.0.1
langchain_openai~=0.2.12
langchain-community~=0.3.11